Re-implement the pipeline using the production Dockerfile
- Switches from using the ruby agent to using the actual Dockerfile
which will be used in production. This offers a number of benefits,
notably more accuracy (the image is the same as the one used in prod)
but also better caching (saving ~5m bundle-install time on builds that
don't modify the Gemfile(s)).
- Consequently, implements the Dockerfile for production use. Labels are
now added directly to the Dockerfile, eliminating the need for the
docker-compose.ci.yml file.