Fleshes out the dev stack
Implements a variety of Compose and Jenkins changes discussed at the
04/24/20 Rails development meeting in LIT, including:
- Adds a working docker-compose.yml configuration. This includes the
site itself (aliased "web") along with a MariaDB database.
- In the Dockerfile, delays the --deployment installation until the
"production" target. This allows you to bundle-add and bundle-install in
the "development" target without having to disable deployment mode,
while still verifying the Gemfile.lock in the production target (where,
to prevent it from calling out to the internet a second time, we also
pass the --local flag).
- Consolidates all testing tasks into a single rake command. If the "CI"
environment variable is set, then rake executes setup, testing, and
teardown tasks, simplifying the Jenkinsfile to a single "Test" stage
that runs specs, brakeman, and bundler-audit.