Bug mitigation: Use timecop to fix item test times
The items_controller does not properly handle timezones and DST, leading to test failures within 24h of the onset of daylight savings time. This PR temporarily works around that bug by using the timecop gem to fix time. To simplify re-surfacing this bug in the future, the environment variable RSPEC_FORCE_DST_FAILURE can be set to a bool-truth value ("true", "t", etc.) to set the test to run at a known buggy time.