app: Don't disable submit button on submission
While it is a good idea to prevent multiple submissions of the same
content, disabling the submit button has a lot of negatives:

* If the user's Internet connection is interrupted, and they click
  Submit, the user-agent will likely render a "Web page not available"
  message.  When they go back, the user-agent may 'remember' the
  button being disabled and prevent the form from being submitted.

* Screen readers cannot communicate the change in form availability
  without significantly more work (at least an aria-live container
  noting that the form has been submitted and cannot be interacted
  with any longer until the results are available).

* For the search form, if the user notices a typo and tries to fix
  it before the results page loads, they cannot submit the fixed query.

This commit removes the `data-disable-with` property from the five
button elements that had it, thereby not disabling the button on form
submission.  Solving the duplicate submission issue is outside the
scope of this ticket.

Closes: ADA-396